Pass a custom login response back to the user
This allows the login failure (or success) to be reported
along with a meaningful message to the user.

At Fastmail we use this for things like "FTP service is not
available on your account" or "you have too many currently
active connections" rather than a generic password failure.

We also report "you have a read-only connection" for users
who have configured a separate read-only password, and we
report one-time password usage if they have used a one-time
password.

USAGE: return a second paramater - the string to show - from
the authentication_hook.